+++ packages/liveblocks-core/src/protocol/UrlMetadata.ts
@@ -1,0 +1,6 @@
+export type UrlMetadata = {
+  title?: string;
+  description?: string;
+  image?: string;
+  icon?: string;
+};