+++ packages/liveblocks-react-ui/src/components/internal/AiComposer.tsx
@@ -1,13 +1,9 @@
 import {
   type AiChatMessage,
-  type AiKnowledgeSource,
   type CopilotId,
   type MessageId,
 } from "@liveblocks/core";
-import {
-  useSendAiMessage,
-  type UseSendAiMessageOptions,
-} from "@liveblocks/react";
+import { useSendAiMessage } from "@liveblocks/react";
 import {
   type ComponentProps,
   type FormEvent,
@@ -92,11 +88,6 @@ export interface AiComposerProps
   /**
    * @internal
    */
-  knowledge?: AiKnowledgeSource[];
-
-  /**
-   * @internal
-   */
   branchId?: MessageId;
 
   /**
@@ -160,7 +151,6 @@ export const AiComposer = forwardRef<HTMLFormElement, AiComposerProps>(
       overrides,
       className,
       chatId,
-      knowledge: localKnowledge,
       branchId,
       copilotId,
       responseTimeout,
@@ -175,11 +165,7 @@ export const AiComposer = forwardRef<HTMLFormElement, AiComposerProps>(
       stream,
       copilotId,
       timeout: responseTimeout,
-      // TODO: We shouldn't need to pass knowledge from AiChat to AiComposer
-      //       to useSendAiMessage, ideally it would be attached to a chat ID
-      //       behind the scenes inside AiChat.
-      knowledge: localKnowledge,
-    } as UseSendAiMessageOptions);
+    });
 
     const handleComposerSubmit = useCallback(
       (message: AiComposerSubmitMessage, event: FormEvent<HTMLFormElement>) => {