+++ packages/liveblocks-react/test-d/no-augmentation.test-d.tsx
@@ -984,13 +984,13 @@ import { expectAssignable, expectError, expectType } from "tsd";
       expectType<"tool-invocation">(status.partType);
       expectType<string>(status.toolName);
     } else {
-      expectType<"text" | "reasoning" | "retrieval" | undefined>(
+      expectType<"text" | "reasoning" | "retrieval" | "sources" | undefined>(
         status.partType
       );
       expectType<undefined>(status.toolName);
     }
   } else {
-    expectType<"loading" | "idle">(status.status);
+    expectType<"disconnected" | "loading" | "idle">(status.status);
     expectType<undefined>(status.partType);
     expectType<undefined>(status.toolName);
   }
@@ -1006,13 +1006,13 @@ import { expectAssignable, expectError, expectType } from "tsd";
       expectType<"tool-invocation">(status.partType);
       expectType<string>(status.toolName);
     } else {
-      expectType<"text" | "reasoning" | "retrieval" | undefined>(
+      expectType<"text" | "reasoning" | "retrieval" | "sources" | undefined>(
         status.partType
       );
       expectType<undefined>(status.toolName);
     }
   } else {
-    expectType<"loading" | "idle">(status.status);
+    expectType<"disconnected" | "loading" | "idle">(status.status);
     expectType<undefined>(status.partType);
     expectType<undefined>(status.toolName);
   }
@@ -1027,13 +1027,13 @@ import { expectAssignable, expectError, expectType } from "tsd";
       expectType<"tool-invocation">(status.partType);
       expectType<string>(status.toolName);
     } else {
-      expectType<"text" | "reasoning" | "retrieval" | undefined>(
+      expectType<"text" | "reasoning" | "retrieval" | "sources" | undefined>(
         status.partType
       );
       expectType<undefined>(status.toolName);
     }
   } else {
-    expectType<"loading" | "idle">(status.status);
+    expectType<"disconnected" | "loading" | "idle">(status.status);
     expectType<undefined>(status.partType);
     expectType<undefined>(status.toolName);
   }
@@ -1048,13 +1048,13 @@ import { expectAssignable, expectError, expectType } from "tsd";
       expectType<"tool-invocation">(status.partType);
       expectType<string>(status.toolName);
     } else {
-      expectType<"text" | "reasoning" | "retrieval" | undefined>(
+      expectType<"text" | "reasoning" | "retrieval" | "sources" | undefined>(
         status.partType
       );
       expectType<undefined>(status.toolName);
     }
   } else {
-    expectType<"loading" | "idle">(status.status);
+    expectType<"disconnected" | "loading" | "idle">(status.status);
     expectType<undefined>(status.partType);
     expectType<undefined>(status.toolName);
   }